feat: ver1.6

+++ factory/Discriminator.py
@@ -6,10 +6,10 @@ class Discriminator(nn.Module):
         super(Discriminator, self).__init__()
         self.conv1 = nn.Conv1d(crop_len, 2 * dim_neck, 3)
         self.conv2 = nn.Conv1d(2 * dim_neck, dim_neck, 3)
-        self.conv3 = nn.Conv1d(dim_neck, dim_neck / 2, 3)
+        self.conv3 = nn.Conv1d(dim_neck, int(dim_neck / 2), 3)
         self.leaky_relu = nn.LeakyReLU()
         self.bn1 = nn.BatchNorm1d(dim_neck)
-        self.bn2 = nn.BatchNorm1d(dim_neck / 2)
+        self.bn2 = nn.BatchNorm1d(int(dim_neck / 2))
         self.flatten = nn.Flatten()
         # TODO: parms
         self.dense1 = nn.Linear(1628, 1)
